PV:3. RPC3 Client.

+++ trunk/client/FMP/functions.cpp
@@ -81,7 +81,7 @@ float GetDist(float x1,float y1,float z1,float x2, float y2, float z2)
 	return sqrt( (x1-x2)*(x1-x2) + (y1-y2)*(y1-y2) + (z1-z2)*(z1-z2) );
 }
 
-int _GetClosestCar(float x, float y, float z, float r)
+short _GetClosestCar(float x, float y, float z, float r)
 {
 	float min = r;
 	int id = -1;
@@ -100,7 +100,7 @@ int _GetClosestCar(float x, float y, float z, float r)
 	return id;
 }
 
-int GetPlayerCar(Vehicle car)
+short GetPlayerCar(Vehicle car)
 {
 	for(int i = 0; i < MAX_CARS; i++)
 	{
