+++ trunk/client/FMP/functions.h
@@ -4,5 +4,5 @@
 void GetTarget(float ax, float az, float *x, float *y, float *z, float mn);
 void GetCamTargetedCoords(float *x, float *y, float *z);
 float GetDist(float x1,float y1,float z1,float x2, float y2, float z2);
-int _GetClosestCar(float x, float y, float z, float r);
-int GetPlayerCar(Vehicle);
+short _GetClosestCar(float x, float y, float z, float r);
+short GetPlayerCar(Vehicle);
